+++ src/main/java/de/mari_023/ae2wtlib/wat/WATScreen.java
@@ -5,11 +5,10 @@ import net.minecraft.world.entity.player.Inventory;
 
 import appeng.client.gui.me.patternaccess.PatternAccessTermScreen;
 import appeng.client.gui.style.ScreenStyle;
-import appeng.client.gui.widgets.BackgroundPanel;
 import appeng.client.gui.widgets.ToolboxPanel;
-import appeng.client.gui.widgets.UpgradesPanel;
 import appeng.menu.SlotSemantics;
 
+import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 import de.mari_023.ae2wtlib.wut.IUniversalTerminalCapable;
 
@@ -19,10 +18,11 @@ public class WATScreen extends PatternAccessTermScreen<WATMenu> implements IUniv
         if (getMenu().isWUT())
             addToLeftToolbar(cycleTerminalButton());
 
-        widgets.add("upgrades", new UpgradesPanel(getMenu().getSlots(SlotSemantics.UPGRADE), getMenu().getHost()));
+        widgets.add("scrollingUpgrades",
+                new ScrollingUpgradesPanel(menu.getSlots(SlotSemantics.UPGRADE), menu.getHost(), widgets));
         if (getMenu().getToolbox().isPresent())
             widgets.add("toolbox", new ToolboxPanel(style, getMenu().getToolbox().getName()));
-        widgets.add("singularityBackground", new BackgroundPanel(style.getImage("singularityBackground")));
+        addSingularityPanel(widgets, getMenu());
     }
 
     @Override
@@ -31,8 +31,7 @@ public class WATScreen extends PatternAccessTermScreen<WATMenu> implements IUniv
     }
 
     @Override
-    public void storeState() {
-    }
+    public void storeState() {}
 
     @Override
     public boolean keyPressed(int keyCode, int scanCode, int keyPressed) {