+++ src/main/java/de/mari_023/ae2wtlib/wet/WETScreen.java
@@ -5,8 +5,9 @@ import net.minecraft.world.entity.player.Inventory;
 
 import appeng.client.gui.me.items.PatternEncodingTermScreen;
 import appeng.client.gui.style.ScreenStyle;
-import appeng.client.gui.widgets.BackgroundPanel;
+import appeng.menu.SlotSemantics;
 
+import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 import de.mari_023.ae2wtlib.wut.IUniversalTerminalCapable;
 
@@ -15,7 +16,9 @@ public class WETScreen extends PatternEncodingTermScreen<WETMenu> implements IUn
         super(container, playerInventory, title, style);
         if (getMenu().isWUT())
             addToLeftToolbar(cycleTerminalButton());
-        widgets.add("singularityBackground", new BackgroundPanel(style.getImage("singularityBackground")));
+        addSingularityPanel(widgets, getMenu());
+        widgets.add("scrollingUpgrades",
+                new ScrollingUpgradesPanel(menu.getSlots(SlotSemantics.UPGRADE), menu.getHost(), widgets));
     }
 
     @Override